getCaseNoteAction -> getJudgesCaseNoteAction
